Move chat UI files up one directory
