+++ src/features/chat/NavigationMarker.tsx
@@ -8,7 +8,7 @@ import {
   GitBranch
 } from 'lucide-react'
 import { NavigationItem } from '@/shared/types'
-import { useChatStore } from '../state'
+import { useChatStore } from './state'
 import { useNavigationStore } from '@/features/navigation/state'
 
 interface NavigationMarkerProps {
